Templatize factories
Right now, users need to create a new derived factory class for
forms/controls any time they want to make a change, including where a
control gets placed on a form. If there are two of the same type of
control on a form, each needs its own factory.

The real goal of the factories is to instantiate a derived form/control.
This same goal can be achieved using templated factories.

This commit templatizes the factories to allow typical code to be easier
to read & write.