Fix 64-bit builds
There was a size_t being passed to a function that took an int.
This produced a warning about possible truncation, which was promoted to
an error.

Right now, we blindly cast it to an int. Later, we should check that it
actually fits inside an int. I have left a TODO for this.