Add styling with example
Certain controls can have styles applied. For example, a textbox can
have a password style which would not display the characters entered.

This commit adds styles along with an example that illustrates the
concept.

This commit also reorders the styling example to be #2, moving the
control gallery example to #3. This is because it made more sense for
the control gallery to use styles.