Don't expose HWND to form ctor
There was a TODO to not pass the HWND to the form's ctor.
This exposed a platform-specific piece. To be cross-platform, this would
need to be fixed.

This commit fixes it by no longer relying on an inheritance hierarchy.
Instead, type erasure is used to route messages.

This also means default behavior (like OnClose calling
PostExitMessage(0)) no longer happens for free. As a result, that
default behavior was removed.

There are some knock-on effects of this change like cleaning the
FormAllocator.