Improve /Wall handling
MSVC's /Wall is not great. A lot of standard library & Windows headers
will generate warnings. So it should not be used for regular dev, in my
opinion.

However, it is nice to check in on it from time to time to see if
anything might be unintentionally risky.

This commit addresses all the reasonable issues that /Wall showed.
Some warnings were left as they're basically useless (such as a warning
that padding bytes will be added to the end of a class).