Fix up to warning level 4 on MSVC
MSVC warns with C26813 because a bitmask isn't being used in the form
"if (my_flag & FLAG_VALUE_1)". But it cannot be used that way because it
is an enum class.

This commit disables C26813 only at that location since it is an
incorrect warning. A bug has also been filed against MSVC.