Fix example 1 on Linux
Previously, example 1's text box didn't resize with the form.
Additionally, the form's title was not set and the form size was fixed.

Now, the form's style is respected (allowing resizing), the custom form
receives a resize notification, and controls can be resized. Additionally,
the form's title is now set.

This bring Linux example 1 to parity with Windows.