Add dark mode support
Dark mode support in Win32 is spotty. There is a function to update the
title bar. And there is a hidden registry value to check if dark mode is
enabled. Also, there is a WM_SETTINGCHANGED sent when the mode is
changed.

With these things, a program can support dark mode.

However, the common controls do not have a way of painting in dark mode.
They need to be painted manually.

If a user is painting their own controls anyway, perhaps this is fine.
And even if not, this is perhaps better than nothing.

This commit adds support for dark mode.