Rename the entry point (#78)
Previously, the entry point file was named EntryPoint.hpp/.cpp.
If other projects used that same name, there could be a conflict in
intermediate files. In fact, this is pretty much the default.

In order to prevent that conflict, the files are renamed to
maxGUIEntryPoint.hpp/.cpp. This is far less likely to cause a conflict.